requirements.txt
argparse
chromadb
datetime
dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapiclient
langchain
langchain-community
langchain-openai
langchain-text-splitters
logging
openai
pickle
pypdf
python-dotenv
streamlit
supabase
typing
google-drive-api